+++ src/motors.c
@@ -2,13 +2,13 @@
 #include "mcu.h"
 #include <stdlib.h>
 
-static int16_t ConstrainPwm(int16_t pwm)
+static int ConstrainPwm(int pwm)
 {
     return pwm > MOTOR_PWM_MAX ? MOTOR_PWM_MAX :
             pwm < -MOTOR_PWM_MAX ? -MOTOR_PWM_MAX : pwm;
 }
 
-static void SetLeftPwm(int16_t pwm)
+static void SetLeftPwm(int pwm)
 {
     pwm = ConstrainPwm(pwm);
     if (pwm > 0) {
@@ -20,7 +20,7 @@ static void SetLeftPwm(int16_t pwm)
     }
 }
 
-static void SetRightPwm(int16_t pwm)
+static void SetRightPwm(int pwm)
 {
     pwm = ConstrainPwm(pwm);
     if (pwm > 0) {
@@ -32,16 +32,16 @@ static void SetRightPwm(int16_t pwm)
     }
 }
 
-void Motors_SetPwm(int16_t left, int16_t right)
+void Motors_SetPwm(int left, int right)
 {
     SetLeftPwm(left);
     SetRightPwm(right);
 }
 
-void Motors_GetPwm(int16_t *left, int16_t *right)
+void Motors_GetPwm(int *left, int *right)
 {
-    *left = (int16_t)TIM_GetCapture2(TIM4) - (int16_t)TIM_GetCapture1(TIM4);
-    *right = (int16_t)TIM_GetCapture3(TIM4) - (int16_t)TIM_GetCapture4(TIM4);
+    *left = (int)TIM_GetCapture2(TIM4) - (int)TIM_GetCapture1(TIM4);
+    *right = (int)TIM_GetCapture3(TIM4) - (int)TIM_GetCapture4(TIM4);
 }
 
 static int execute(int argc, char *argv[])
@@ -49,8 +49,8 @@ static int execute(int argc, char *argv[])
     if (argc != 2)
         return -1;
 
-    int16_t pwmLeft = atoi(argv[0]);
-    int16_t pwmRight = atoi(argv[1]);
+    int pwmLeft = atoi(argv[0]);
+    int pwmRight = atoi(argv[1]);
 
     Motors_SetPwm(pwmLeft, pwmRight);
     return 0;
@@ -58,7 +58,7 @@ static int execute(int argc, char *argv[])
 
 static void WriteTelemetry(char out[TELEMETRY_STRING_SIZE])
 {
-    int16_t pwmLeft, pwmRight;
+    int pwmLeft, pwmRight;
 
     Motors_GetPwm(&pwmLeft, &pwmRight);
     snprintf(out, TELEMETRY_STRING_SIZE, "L:%d\tR:%d\n", pwmLeft, pwmRight);
